refactor(engine): extract KVCacheInitializer (init_kv_cache) to its own TU
Moves Engine::init_kv_cache (~255 LOC) to src/runtime/engine_kv_cache_init.cpp.

Declaration in engine.h unchanged. Body byte-identical.

Phase 4 of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
